Sample for DE vehicleignorelist.json
All vehicles were always added to the missions. Where it is missing, it is gradually added.
So that new players or those who do not want to have all the vehicles have the opportunity to easily take these vehicles out again without touching the missions. This should only be used as a template for copying and pasting to these.